Make block_memory size configurable.

+++ include/bitcoin/node/block_arena.hpp
@@ -47,13 +47,13 @@ private:
     bool do_is_equal(const arena& other) const NOEXCEPT override;
 
     // These are thread safe.
-    const uint64_t capacity_;
+    const size_t capacity_;
     std::shared_mutex field_mutex_{};
     std::shared_mutex remap_mutex_{};
 
     // These are protected by mutex.
     uint8_t* memory_map_;
-    uint64_t offset_{};
+    size_t offset_{};
 
 };
